+++ extensions/graph-ui-temp/ui/graph-page.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export function GraphPage() {
+  return <div>WIP - initial graph page</div>;
+}
